Always minify preamble script.
We don't know the exact size at which things break, so don't try to be
smart about it.

Thinking about it, this whole approach isn't appropriate to a generic
tool and is going to break horribly for (e.g.) Windows VMs where the
preamble is unlikely to be a bash script. We should probably remove it
completely and leave it up to the caller to Do The Right Thing.