Don't specify gem versions
Don't specify the versions that this meta-gem installs.

By not specifying the versions, new users of vCloud Tools will install
the latest available versions of the tools and we avoid having to
increment the versions each time a gem is updated.

* * *

Also bump the version to 0.9.0 and update the CHANGELOG accordingly.